feat(bat): senv set GOBIN last in `PATH`, mods settings
If we want the wrapper mods.bat to take precedence over the go-installed
`GOBIN\mods.exe`, we need GOBIN to be last in the `PATH`, after `HOME/bin`
Include a default model and example prompt within the mods.bat script itself.
This helps new users understand the script's functionality and potential configurations.

- Added default model setting within the `mods.bat` script.
- Included an example prompt for the `mods.bat` script.
- Modified `senv.bat` to correctly prepend GOROOT to PATH.