fix(setup.bat): avoid sysinternals double installation
The `sysinternals` option name in the setup script was incorrect.
This patch corrects the option name to `sysinternalsSuites` to reflect the actual
suite name PRGS folder installation.

If the name was incorrect, re-declaring sysinternalsSuites in in a personal
install_home.list made sysinternalsSuites to be installed a second time, since
it was already installed as a 'global' tool (a tool installed for everybody, not
matter their profile.

What:
- Changed the `sysinternals` option name to `sysinternalsSuites` in the setup.bat script.