fix(setup.bat): Reset fn on a 'system' install
If a prg/tool is marked as system, fn is not computed, make sure it is reset, or
it would still have the filename of the previous tool installed.

Note: a 'p' (pattern) 'system' install means one where we need to search for the
Windows system installation path, instead of the regular `%PRGS%/mytool` path.

What:
- Added explicit setting of `fname` variable to empty string when `p` is "system".