feat(bin): detect wildfly.bat action to perform
Why:
The existing `wildfly.bat` script only checked the server status.
This change adds the ability to detect the last odd parameter as the action to
execute on the Wildfly server from the Windows command line.
This makes it simpler to manage the Wildfly server remotely and avoids the need
for GUI interaction if run from a system account.

What:
- Added `WF_ACTION` variable to control the script's action.
- Implemented `:status`, and `:run` functions.
- Modified `:init_params` to parse the `WF_ACTION` argument if provided.
  Default: `status`.
- Added logic to call the appropriate action based on `WF_ACTION`.
- Added logic to prevent accidental container re-creation.
- Improved output messages to be more informative.