fix(bin): wildfly persists JAVA_HOME across calls
Why: Persisting JAVA_HOME ensures that subsequent calls to the script can
reliably access the correct Java environment.
Previously, the JAVA_HOME variable set by the function `switchjdk` got lost when
`switchjdk` terminated.
The loss of this information interfered with the correct setup and execution of
the Java application. This change addresses this issue and improves the reliability of the system.

What:
- Added `JAVA_HOME` to the list of persisted environment variables.
- Set default management user to `%USERNAME%-adm`.
- Set default management password to `%USERNAME%-adm-pwd`.
- Improved reporting using `echos.bat`.
- Explicitly set JAVA_HOME to jdk path.