fix(bin): gsh.bat prompt must not depend on mods.exe
Why
The "prompt" mode can still try and seek a prompts from the `mods.exe` settings.
But when said `mods.exe` is /not/ installed, we must generate a prompt locally.
Previously, the script would fail in these situations.
The updated script now provides more informative warning messages and creates a
default prompt if necessary.

Reminder: the prompt mod is made to dump a prompt + a diff to the user clipboard,
for said user to paste it to any LLM web page of their chosing.

What
- Added error handling for missing `mods.exe` by checking its availability.
- Added more informative warning messages if `mods.exe`, `mods.yml`,
  or `config_path` is not found.
- Implemented a fallback mechanism to create a default prompt from within
  `gsh.bat` if `mods.yml` is not found,
- Added space around comparison operator in `if "%~1"=="prompt"`.
- Added `del tmp.txt 2>NUL` before `where mods.exe`, to ensure its creation.
- Use explicitly `mods.exe --dirs`, not just `mods --dirs`.