fix(bin): `mods.bat` Correctly set `EDITOR` to npp
Why: The previous command was not working as intended because the path to
notepad++ (and its `settings` folder) was not properly wrapped with double-quotes.
Sidestep the issue by changing the execution path to the Notepad++ one.
No more path with spaces, since `EDITOR` references a relative path to
`notepad++.exe` now.

What:
- Changed the `EDITOR` variable to point a relative path to notepad++.
- Changed the settings directory to `settings` relative path.
- Added `pushd` and `popd` commands to manage the relative paths.