refactor(installs): Migrate drawio to post install
Why: The previous installation method of Drawio was not working (calling the
`draw.io-26.1.1-windows-no-installer.exe` directly. It needs to be unziped, and
(in the post step, the $PLUGINSDIR\app-64.7z also needs to be unziped.
That produced the `draw.io.exe` with all the necessary DLL, and avoids the issue
https://github.com/jgraph/drawio-desktop/issues/2039

What:
- Created `drawios.post.bat` to handle drawio installation.
- Created `drawios.alias.bat` to handle draw.io alias creation.
- Removed `drawios.install.bat`.
- Updated `inst_prg.bat` to call post and alias install scripts.
- Added call DOSKEY at the end of `inst_prg.bat`.